+++ eemeter/caltrack.py
@@ -1,10 +1,10 @@
 from collections import Counter
+import traceback
 
 import numpy as np
 import pandas as pd
 import pytz
 import statsmodels.formula.api as smf
-import traceback
 
 from .api import (
     CandidateModel,
@@ -14,10 +14,16 @@ from .api import (
     ModelPrediction,
 )
 from .exceptions import MissingModelParameterError, UnrecognizedModelTypeError
-from .transform import day_counts, overwrite_partial_rows_with_nan
+from .features import (
+    compute_time_features,
+    compute_temperature_features,
+    compute_temperature_bin_features,
+    compute_occupancy_feature,
+    merge_features,
+)
 from .metrics import ModelMetrics
-
-from .features import compute_temperature_features
+from .segmentation import SegmentModel
+from .transform import day_counts, overwrite_partial_rows_with_nan
 
 
 __all__ = (
@@ -37,6 +43,9 @@ __all__ = (
     "get_hdd_only_candidate_models",
     "get_cdd_hdd_candidate_models",
     "select_best_candidate",
+    "caltrack_hourly_fit_feature_processor",
+    "caltrack_hourly_prediction_feature_processor",
+    "fit_caltrack_hourly_model_segment",
 )
 
 
@@ -1756,3 +1765,86 @@ def plot_caltrack_candidate(
         ax.set_title(title)
 
     return ax
+
+
+def caltrack_hourly_fit_feature_processor(
+    segment_name, segmented_data, occupancy_lookup, temperature_bins
+):
+    # get occupied feature
+    hour_of_week = segmented_data.hour_of_week
+    occupancy = occupancy_lookup[segment_name]
+    occupancy_feature = compute_occupancy_feature(hour_of_week, occupancy)
+
+    # get temperature bin features
+    temperatures = segmented_data.temperature_mean
+    bin_endpoints_list = (
+        temperature_bins[segment_name].index[temperature_bins[segment_name]].tolist()
+    )
+    # TODO(philngo): combine with compute_temperature_features
+    temperature_bin_features = compute_temperature_bin_features(
+        segmented_data.temperature_mean, bin_endpoints_list
+    )
+
+    # combine features
+    return merge_features(
+        [
+            segmented_data[["meter_value", "hour_of_week"]],
+            occupancy_feature,
+            temperature_bin_features,
+            segmented_data.weight,
+        ]
+    )
+
+
+def caltrack_hourly_prediction_feature_processor(
+    segment_name, segmented_data, occupancy_lookup, temperature_bins
+):
+    # hour of week feature
+    hour_of_week_feature = compute_time_features(segmented_data.index)
+
+    # occupancy feature
+    occupancy = occupancy_lookup[segment_name]
+    occupancy_feature = compute_occupancy_feature(
+        hour_of_week_feature.hour_of_week, occupancy
+    )
+
+    # get temperature bin features
+    temperatures = segmented_data
+    bin_endpoints_list = (
+        temperature_bins[segment_name].index[temperature_bins[segment_name]].tolist()
+    )
+    # TODO(philngo): combine with compute_temperature_features
+    temperature_bin_features = compute_temperature_bin_features(
+        segmented_data.temperature_mean, bin_endpoints_list
+    )
+
+    # combine features
+    return merge_features(
+        [
+            hour_of_week_feature,
+            occupancy_feature,
+            temperature_bin_features,
+            segmented_data.weight,
+        ]
+    )
+
+
+def get_hourly_model_formula(data):
+    bin_occupancy_interactions = "".join(
+        [" + {}:occupancy".format(c) for c in data.columns if "bin" in c]
+    )
+    return "meter_value ~ C(hour_of_week) - 1{}".format(bin_occupancy_interactions)
+
+
+def fit_caltrack_hourly_model_segment(segment_name, segment_data):
+    formula = get_hourly_model_formula(segment_data)
+    model = smf.wls(formula=formula, data=segment_data, weights=segment_data.weight)
+    model_params = {coeff: value for coeff, value in model.fit().params.items()}
+    warnings = []
+    return SegmentModel(
+        segment_name=segment_name,
+        model=model,
+        formula=formula,
+        model_params=model_params,
+        warnings=warnings,
+    )
